requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
argcomplete==3.6.1
bidict==0.23.1
blinker==1.8.2
boto3==1.37.20
botocore==1.37.20
catkin-pkg==1.0.0
certifi==2025.1.31
cfn-flip==1.3.0
charset-normalizer==3.4.1
click==8.1.8
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.19.0
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.10
contourpy==1.1.1
coverage==7.6.1
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
dnspython==2.6.1
docutils==0.20.1
durationpy==0.9
email_validator==2.2.0
empy==4.2
eventlet==0.39.1
exceptiongroup==1.2.2
flake8==7.1.2
flake8-blind-except==0.2.1
flake8-builtins==2.5.0
flake8-class-newline==1.6.0
flake8-comprehensions==3.15.0
flake8-deprecated==2.2.1
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
flake8-quotes==3.4.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.57.0
greenlet==3.1.1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
kappa==0.6.0
kiwisolver==1.4.7
Mako==1.3.9
MarkupSafe==2.1.5
matplotlib==3.7.5
mccabe==0.7.0
mock==5.2.0
mypy==0.931
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
numpy==1.24.4
openai==1.69.0
packaging==24.2
pandas==2.0.3
pep8==1.7.1
pillow==10.4.0
placebo==0.9.0
playsound==1.3.0
pluggy==1.5.0
psutil==7.0.0
pycaw==20240210
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pydocstyle==6.3.0
pydub==0.25.1
pyflakes==3.2.0
PyMySQL==1.1.1
pyparsing==3.1.4
pyreadline3==3.5.4
pytest==8.3.5
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-repeat==0.9.3
pytest-rerunfailures==14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-slugify==8.0.4
python-socketio==5.12.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
SpeechRecognition==3.10.4
SQLAlchemy==2.0.39
text-unidecode==1.3
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
troposphere==4.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
vcstool==0.3.0
waitress==3.0.0
Werkzeug==3.0.6
wsproto==1.2.0
WTForms==3.1.2
zappa==0.59.0
zipp==3.20.2
